Add a first, simple version, of an SQL INSERT
`Insert` represents an SQL INSERT, and `InsertStatement` the actual
statement.

The table schema can be defined by extending `Table`, although only
string, and int types are supported for the moment.

Insertion works directly with the `PreparedStatement`, but that will
probably change in the future to precisely track expressions
used in a statement, in order to read/write columns without ambiguity.