Add a first version, of an SQL UPDATE
`Update` represents an SQL UPDATE, and `UpdateStatement` the actual
statement. No support for WHERE clause yet.

`Dataset` is now either closed - with a fixed set of columns,
typically for inserts or complete updates, or opened, e.g. for
partial updates. When the set is opened, the columns returned
are the set columns in order.