Use TIMESTAMP WITH TIMEZONE for Instant type columns
Change SQL type for Instant columns to TIMESTAMP WITH TIMEZONE
using the UTC timezone. This better fits the fact that a
timestamp is a point on the timeline. Some drivers (e.g. Postgres)
always store timestamps in UTC and getting the timestamp
in the client timezone (see JDBC spec) is confusing.

TIMESTAMP (WITHOUT TIMEZONE) is reserved for LocalDate.